requirements.txt
langchain==0.2.13
langchain_experimental==0.0.64
langchain-openai==0.1.25
pandas==2.2.3
numpy==1.26.4
imbalanced-learn==0.12.3
matplotlib==3.9.2
geopy==2.4.1
geojson==3.1.0
shapely==2.0.6
streamlit==1.39.0
openai== 1.50.2
llama-index==0.11.15
llama-index-llms-openai==0.2.10
nltk==3.9.1
pypdf==4.3.1
streamlit-chat==0.1.1
python-dotenv==1.0.1